@ElvishArtisan 2015-03-10
On Mar 10, 2015, at 04:39 35, Alban Peignier <notifications@github.com> wrote:

> Any chance for v2.11.0 ? :-)

I’d say, chances are excellent for both!  :)

Cheers!


|----------------------------------------------------------------------|
| Frederick F. Gleason, Jr. |              Chief Developer             |
|                           |              Paravel Systems             |
|----------------------------------------------------------------------|
|          A room without books is like a body without a soul.         |
|                                         -- Cicero                    |
|----------------------------------------------------------------------|