@albanpeignier 2015-03-20
You can test this pull request by using the debian packages we've built. They include Rivendell 2.10.3 and the patch proposed by this pull request : 

```
wget http://download.tryphon.eu/debian/wheezy/librivendell_2.10.3-1+pr44_amd64.deb \
  http://download.tryphon.eu/debian/wheezy/rivendell_2.10.3-1+pr44_amd64.deb \
  http://download.tryphon.eu/debian/wheezy/rivendell-server_2.10.3-1+pr44_amd64.deb

sudo dpkg -i librivendell_2.10.3-1+pr44_amd64.deb rivendell_2.10.3-1+pr44_amd64.deb rivendell-server_2.10.3-1+pr44_amd64.deb
```

caed (especially with large sound cards) should use half the processor.

Packages are available for wheezy amd64 but we can build any debian/ubuntu platform.