@ElvishArtisan 2015-07-31
Merged to master.  Thanks!